@@ -1,3 +1,3 @@
-export const ALGOLIA_APP_ID = process.env.REACT_APP_ALGOLIA_APP_ID;
-export const ALGOLIA_API_KEY = process.env.REACT_APP_ALGOLIA_API_KEY;
-export const KAPPA_API_URL = process.env.REACT_APP_KAPPA_API_URL;
+export const ALGOLIA_APP_ID = import.meta.env.VITE_ALGOLIA_APP_ID;
+export const ALGOLIA_API_KEY = import.meta.env.VITE_ALGOLIA_API_KEY;
+export const KAPPA_API_URL = import.meta.env.VITE_KAPPA_API_URL;
